Fix binary assigns
ZEND_ASSIGN_ADD and co. were changed to ZEND_ASSIGN_OP, with ZEND_ADD stored in
extended_value. This commit makes some small changes to handle that so the
correct functions can still be called.

I haven't found a way to get __assign to work. It seems like op_code is not
longer called at all when doing something like `$a = 2;`. I've removed it from
the tests for now, I think it's a PHP thing.